+++ app/integrations/kafka/kafka_connection.py
@@ -2,7 +2,7 @@ import asyncio
 import json
 import uuid
 
-from app.kafka.kafka_client import get_producer, get_consumer
+from app.integrations.kafka.kafka_client import get_producer, get_consumer
 from app.config import KAFKA_BOOTSTRAP_SERVER
 from app.core.ai_logger import get_ai_logger
