+++ app/services/v1/plan_writer.py
@@ -1,4 +1,4 @@
-from app.schemas.v1.group_writer import GroupGenerationRequest
+from app.schemas.groups.group_writer import GroupGenerationRequest
 from app.core.ai_logger import get_ai_logger
 from app.core.vertex_client import smart_generate
 from app.prompts.prompt_loader import load_prompt_template
